+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Unreleased
 ### Changes
 
 - Migrate `vm2` to `node:vm`
+- "makecomapp.json not found" error message improvement
+- "Workspace" word change to "Folder"
 
 1.3.31 (pre-release)
 --------------------